+++ moPepGen/cli/parse_star_fusion.py
@@ -95,7 +95,9 @@ def parse_star_fusion(args:argparse.Namespace) -> None:
 
     common.print_start_message(args)
 
-    genome, anno, *_ = common.load_references(args, load_canonical_peptides=False)
+    ref_data = common.load_references(args, load_canonical_peptides=False)
+    genome = ref_data.genome
+    anno = ref_data.anno
 
     variants:List[seqvar.VariantRecord] = []