+++ moPepGen/params.py
@@ -73,24 +73,22 @@ class CleavageParams():
             })
         return data
 
-class ReferenceData():
+@dataclasses.dataclass
+class ReferenceData:
     """ Reference related parameters
 
     ## Attributes
         - genome (dna.DNASeqDict)
-        - anno (gtf.GeneAnnotationModel)
+        - anno (GenomicAnnotation)
         - canonical_peptides (Set[str])
         - proteome (aa.AminoAcidSeqDict)
+        - codon_tables (Dict[str, CodonTableInfo])
     """
-    def __init__(self, genome:dna.DNASeqDict, anno:gtf.GenomicAnnotation,
-            canonical_peptides:Set[str], proteome:aa.AminoAcidSeqDict=None,
-            codon_tables:Dict[str,str]=None):
-        """ constructor """
-        self.genome = genome
-        self.anno = anno
-        self.canonical_peptides = canonical_peptides
-        self.proteome = proteome  or aa.AminoAcidSeqDict()
-        self.codon_tables = codon_tables or {}
+    genome: dna.DNASeqDict
+    anno: gtf.GenomicAnnotation
+    canonical_peptides: Set[str] = dataclasses.field(default_factory=set)
+    proteome: aa.AminoAcidSeqDict = dataclasses.field(default_factory=aa.AminoAcidSeqDict)
+    codon_tables: Dict[str, CodonTableInfo] = dataclasses.field(default_factory=dict)
 
 @dataclasses.dataclass
 class CodonTableInfo: